pyproject.toml
[build-system]
requires = ["uv_build >= 0.8.17, <0.9.0"]
build-backend = "uv_build"
[project]
name = "fxml"
version = "0.1.0"
authors = [
  {name = "Chi-Yu Hsieh", email = "chiyuh1996@gmail.com"},
]
description = "Forex Machine Learning Project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "backtesting>=0.6.4",
    "click>=8.2.1",
    "joblib>=1.5.1",
    "jupyter>=1.1.1",
    "lightning>=2.5.2",
    "matplotlib>=3.10.3",
    "mplfinance>=0.12.10b0",
    "numpy>=1.26.0",
    "pandas>=2.3.0",
    "pandas-ta>=0.4.71b0",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.7.0",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.5",
    "ta>=0.11.0",
    "ta-lib>=0.6.4",
    "tensorboard>=2.19.0",
    "torch>=2.7.1",
    "tqdm>=4.67.1",
    "yfinance>=0.2.66",
]
